# docs/swagger.yaml
basePath: /api
definitions:
  main.Area:
    description: A specific area within a room where hedgehogs are housed
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        format: date-time
        type: string
      hedgehogs:
        items:
          $ref: '#/definitions/main.Hedgehog'
        type: array
      height:
        example: 80
        minimum: 1
        type: number
      id:
        example: 1
        type: integer
      max_capacity:
        example: 2
        minimum: 1
        type: integer
      name:
        example: Gabbia 1
        type: string
      room:
        $ref: '#/definitions/main.Room'
      room_id:
        example: 1
        type: integer
      updated_at:
        example: "2024-01-15T10:30:00Z"
        format: date-time
        type: string
      width:
        example: 100
        minimum: 1
        type: number
      x:
        example: 10
        minimum: 0
        type: number
      "y":
        example: 20
        minimum: 0
        type: number
    type: object
  main.Hedgehog:
    description: Information about a hedgehog in the rescue center
    properties:
      area:
        $ref: '#/definitions/main.Area'
      area_id:
        example: 1
        type: integer
      arrival_date:
        example: "2024-01-15T10:30:00Z"
        format: date-time
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        format: date-time
        type: string
      description:
        example: Riccio trovato nel giardino
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Spillo
        type: string
      release_date:
        example: "2024-07-28T10:30:00Z"
        format: date-time
        type: string
      status:
        enum:
        - in_care
        - recovered
        - deceased
        - released
        example: in_care
        type: string
      therapies:
        items:
          $ref: '#/definitions/main.Therapy'
        type: array
      updated_at:
        example: "2024-01-15T10:30:00Z"
        format: date-time
        type: string
      weight_records:
        items:
          $ref: '#/definitions/main.WeightRecord'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing hedgehog rescue center operations
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: La Ninna - Hedgehog Management API
  version: "1.0"
paths:
...
  /areas:
    get:
      consumes:
      - application/json
      description: Get list of all areas with their rooms and hedgehogs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Area'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all areas
      tags:
      - Areas
    post:
      consumes:
      - application/json
      description: Create a new area record
      parameters:
      - description: Area data
        in: body
        name: area
        required: true
        schema:
          $ref: '#/definitions/main.Area'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Area'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new area
      tags:
      - Areas
  /areas/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an area by its ID
      parameters:
      - description: Area ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete area
      tags:
      - Areas
    put:
      consumes:
      - application/json
      description: Update an existing area's information
      parameters:
      - description: Area ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated area data
        in: body
        name: area
        required: true
        schema:
          $ref: '#/definitions/main.Area'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Area'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update area
      tags:
      - Areas
  /export/{dataType}/{format}:
    get:
      consumes:
      - application/json
      description: Export data in various formats (PDF, Excel, CSV)
      parameters:
      - description: Start date filter (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date filter (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Status filter
        in: query
        name: status
        type: string
      - description: Room ID filter
        in: query
        name: room_id
        type: integer
      produces:
      - application/pdf
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      - text/csv
      responses:
        "200":
          description: Exported file
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Quick export data
      tags:
      - Export
  /hedgehogs:
    get:
      consumes:
      - application/json
      description: Get list of all hedgehogs with their areas and relationships
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Hedgehog'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all hedgehogs
      tags:
      - Hedgehogs
    post:
      consumes:
      - application/json
      description: Create a new hedgehog record
      parameters:
      - description: Hedgehog data
        in: body
        name: hedgehog
        required: true
        schema:
          $ref: '#/definitions/main.Hedgehog'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Hedgehog'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new hedgehog
      tags:
      - Hedgehogs
  /hedgehogs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a hedgehog by its ID
      parameters:
      - description: Hedgehog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete hedgehog
      tags:
      - Hedgehogs
    get:
      consumes:
      - application/json
      description: Get a single hedgehog by its ID with related data
      parameters:
      - description: Hedgehog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Hedgehog'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get hedgehog by ID
      tags:
      - Hedgehogs
    put:
      consumes:
      - application/json
      description: Update an existing hedgehog's information
      parameters:
      - description: Hedgehog ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated hedgehog data
        in: body
        name: hedgehog
        required: true
        schema:
          $ref: '#/definitions/main.Hedgehog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Hedgehog'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update hedgehog
      tags:
      - Hedgehogs
...
  /rooms:
    get:
      consumes:
      - application/json
      description: Get list of all rooms with their areas and hedgehogs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Room'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all rooms
      tags:
      - Rooms
    post:
      consumes:
      - application/json
      description: Create a new room record
      parameters:
      - description: Room data
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/main.Room'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Room'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new room
      tags:
      - Rooms
...
  /therapies:
    get:
      consumes:
      - application/json
      description: Get list of all therapies with optional hedgehog_id filter
      parameters:
      - description: Filter by hedgehog ID
        in: query
        name: hedgehog_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Therapy'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all therapies
      tags:
      - Therapies
    post:
      consumes:
      - application/json
      description: Create a new therapy record for a hedgehog
      parameters:
      - description: Therapy data
        in: body
        name: therapy
        required: true
        schema:
          $ref: '#/definitions/main.Therapy'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Therapy'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new therapy
      tags:
      - Therapies
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"